+++ .github/workflows/ci-main.yml
@@ -45,4 +45,4 @@ jobs:
       run: pip install oteltest
 
     - name: Run oteltests
-      run: oteltest tests/ott*
+      run: oteltest tests/integration/*.py
