+++ tests/integration/otel_logging.py
@@ -1,7 +1,7 @@
 import logging
 
 from opentelemetry import trace
-from ott_lib import project_path
+from lib import project_path
 
 MESSAGE = "uh oh!"
 LOGGER_NAME = "logging-ott"
