+++ tests/integration/otel_profile.py
@@ -1,4 +1,4 @@
-from ott_lib import project_path, trace_loop
+from lib import project_path, trace_loop
 
 if __name__ == "__main__":
     trace_loop(12)
