+++ tests/integration/propagator.py
@@ -1,7 +1,7 @@
 from collections.abc import Mapping, Sequence
 from typing import Optional
 
-from ott_lib import UPSTREAM_PRERELEASE_VERSION, project_path
+from lib import UPSTREAM_PRERELEASE_VERSION, project_path
 
 PORT = 8888
