+++ tests/integration/sf_token.py
@@ -1,16 +1,17 @@
-from ott_lib import project_path, trace_loop
+from lib import project_path, trace_loop
 
 if __name__ == "__main__":
-    trace_loop(1)
+    trace_loop(12)
 
 
-class SysMetricsOtelTest:
+class AccessTokenOtelTest:
     def requirements(self):
-        return (project_path(),)
+        return project_path(), "oteltest"
 
     def environment_variables(self):
         return {
             "OTEL_SERVICE_NAME": "my-svc",
+            "SPLUNK_ACCESS_TOKEN": "s3cr3t",
         }
 
     def wrapper_command(self):
@@ -20,9 +21,8 @@ class SysMetricsOtelTest:
         return None
 
     def on_stop(self, telemetry, stdout: str, stderr: str, returncode: int) -> None:
-        from oteltest.telemetry import get_metric_names
-
-        assert "system.cpu.time" in get_metric_names(telemetry)
+        for request in telemetry.get_trace_requests():
+            assert request.headers.get("x-sf-token") == "s3cr3t"
 
     def is_http(self):
         return False
