+++ tests/integration/sysmetrics_enabled.py
@@ -1,4 +1,4 @@
-from ott_lib import project_path, trace_loop
+from lib import project_path, trace_loop
 
 if __name__ == "__main__":
     trace_loop(1)
@@ -11,7 +11,6 @@ class SysMetricsOtelTest:
     def environment_variables(self):
         return {
             "OTEL_SERVICE_NAME": "my-svc",
-            "OTEL_PYTHON_DISABLED_INSTRUMENTATIONS": "system_metrics",
         }
 
     def wrapper_command(self):
@@ -23,7 +22,7 @@ class SysMetricsOtelTest:
     def on_stop(self, telemetry, stdout: str, stderr: str, returncode: int) -> None:
         from oteltest.telemetry import get_metric_names
 
-        assert "system.cpu.time" not in get_metric_names(telemetry)
+        assert "system.cpu.time" in get_metric_names(telemetry)
 
     def is_http(self):
         return False
